+++ app/PersonioIntegration/Widgets/Archive.php
@@ -409,7 +409,7 @@ class Archive extends Widget_Base {
 				'required'      => false,
 			),
 			'showApplicationForm' => array(
-				'label'         => __( 'View application link', 'personio-integration-light' ),
+				'label'         => __( 'View option to apply', 'personio-integration-light' ),
 				'example_value' => 1,
 				'required'      => false,
 			),