+++ blocks/show/src/edit.js
@@ -163,7 +163,7 @@ export default function Edit( object ) {
 						disabled={ disabledFields }
 					/>
 					<ToggleControl
-						label={__('View application link', 'personio-integration-light')}
+						label={__('View option to apply', 'personio-integration-light')}
 						checked={ object.attributes.showApplicationForm }
 						onChange={ value => onChangeApplicationFormVisibility( value, object )  }
 						disabled={ disabledFields }